RecyclerView4 + funcionalidad
Se agrega clicklistener al recyclerview y se agrega la funcionalidad de calcular la conversion al hacer clic en una moneda

+++ app/src/main/java/com/example/conversion/database/MonedaDataBaseDao.kt
@@ -13,7 +13,7 @@ interface MonedaDataBaseDao {
     @Query("select * from table_moneda where nombre= :nom")
     fun getMoneda_nombre(nom: String): Moneda?
     @Query("select * from table_moneda where nro= :nro")
-    fun getMoneda_nro(nro: Int): Moneda?
+    fun getMoneda_nro(nro: Long): Moneda?
 
 //actualizaciones
     @Update
